% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertToExpression.R
\name{convertToExpression}
\alias{convertToExpression}
\title{convertToExpression.}
\usage{
convertToExpression(
  df = expn,
  col_sp_names = "barcode",
  col_gene_names = "gene_id",
  col_counts = "normalized_read_count"
)
}
\arguments{
\item{df}{the data frame do be converted.}

\item{col_sp_names}{the collum with the sample names ids.}

\item{col_gene_names}{the collum with genes names ids.}

\item{col_counts}{the collum with the counts per gene.}
}
\value{
A data frame called df_modified. it contains the data with only sample names, gene names and counts
in the expression format.
Para utilizar: convertToExpression()
}
\description{
The function takes a data frame of rna expression counts and transforms it to
a likely used format required as input for packages like edger and Deseq2.
}
\author{
Bruno Rodrigo Assunção(2024).
}
