% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDEG.R
\name{PlotDEG}
\alias{PlotDEG}
\title{PlotDEG}
\usage{
PlotDEG(
  XData = Edge_LumAvsLumB,
  cutoff_up = 1,
  cutoff_down = -1,
  Title = "Differentially Expressed Genes"
)
}
\arguments{
\item{XData}{the results as showed by edger packages.}

\item{cutoff_up}{the cutoff for up regulated genes.}

\item{cutoff_down}{the cutoff for down regulated genes.}

\item{Title}{the phrase to be exibited as title.}
}
\description{
This function plot Edger DEG results in a plot that takes log Fold Changes
for each genes with different colors for better interpretation.
}
\author{
Bruno Rodrigo Assunção(2024)
}
