% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectFromDataFrame.R
\name{selectFromDataFrame}
\alias{selectFromDataFrame}
\title{selectFromDataFrame}
\usage{
selectFromDataFrame(xdata)
}
\description{
This function is used to takes annotation of another data frame by a shared id/ barcode.
It was used in the DEG-TCGA-WORKFLOW (Bruno R. Assuncao, 2024. \if{html}{\out{<brunor>}} RPubs).
Only for that propose and that use. That need to be modified if for other purpose.
}
